fix relu, sigmoid and argmax
