+++ CMakeLists.txt
@@ -134,14 +134,6 @@ install(TARGETS paddle_enc mpc_data_utils
     LIBRARY DESTINATION ${PADDLE_ENCRYPTED_LIB_PATH}
     LIBRARY DESTINATION ${PADDLE_ENCRYPTED_LIB_PATH})
 
-    install(TARGETS paddle_enc common
-    LIBRARY DESTINATION ${PADDLE_ENCRYPTED_LIB_PATH}
-    LIBRARY DESTINATION ${PADDLE_ENCRYPTED_LIB_PATH})
-
-    install(TARGETS paddle_enc mpc_tensor
-    LIBRARY DESTINATION ${PADDLE_ENCRYPTED_LIB_PATH}
-    LIBRARY DESTINATION ${PADDLE_ENCRYPTED_LIB_PATH})
-
 if (WITH_PSI)
     install(TARGETS psi LIBRARY DESTINATION ${PADDLE_ENCRYPTED_LIB_PATH})
 endif()
