+++ core/privc/CMakeLists.txt
@@ -1,10 +1,8 @@
 set(PRIVC_SRCS
     "privc_context.cc"
-    "integer.cc"
-    "bit.cc"
+    "gc_integer.cc"
+    "gc_bit.cc"
     "ot.cc"
-    "common_utils.cc"
-    #"../paddlefl_mpc/mpc_protocol/context_holder.cc"
 )
 
 add_library(privc_o OBJECT ${PRIVC_SRCS})
@@ -16,8 +14,8 @@ target_link_libraries(privc common mpc_tensor)
 
 cc_test(privc_fixedpoint_tensor_test SRCS fixedpoint_tensor_test.cc DEPS privc)
 cc_test(triplet_generator_test SRCS triplet_generator_test.cc DEPS privc)
-#cc_test(privc_gc_test SRCS privc_gc_test.cc DEPS privc)
-cc_test(privc_bit_test SRCS privc_bit_test.cc DEPS privc)
-cc_test(privc_integer_test SRCS privc_integer_test.cc DEPS privc)
-cc_test(privc_fixedpoint_test SRCS privc_fixedpoint_test.cc DEPS privc)
+
+cc_test(privc_gc_bit_test SRCS privc_gc_bit_test.cc DEPS privc)
+cc_test(privc_gc_integer_test SRCS privc_gc_integer_test.cc DEPS privc)
+cc_test(privc_gc_fixedpoint_test SRCS privc_gc_fixedpoint_test.cc DEPS privc)
