+++ core/privc/common_utils.h
@@ -17,20 +17,70 @@
 #include "core/privc/privc_context.h"
 #include "core/common/tensor_adapter_factory.h"
 #include "core/common/tensor_adapter.h"
-#include "type_utils.h"
 #include "core/paddlefl_mpc/mpc_protocol/context_holder.h"
+#include "core/common/crypto.h"
+#include "core/common/naorpinkas_ot.h"
+#include "core/common/ot_extension.h"
+
+#include "paddle/fluid/platform/enforce.h"
 
 namespace privc {
 
+// use alias name to distinguish TensorAdapter<block>
+// and normal TensorAdapter<int64_t>
+using TensorBlock = common::TensorAdapter<int64_t>;
+
+template<typename T>
+using TensorAdapter = common::TensorAdapter<T>;
+
+using u8 = common::u8;
+using u64 = common::u64;
+
+using block = common::block;
+
+using NaorPinkasOTsender = common::NaorPinkasOTsender;
+using NaorPinkasOTreceiver = common::NaorPinkasOTreceiver;
+
+using AbstractNetwork = paddle::mpc::AbstractNetwork;
+using AbstractContext = paddle::mpc::AbstractContext;
+
 using OT = ObliviousTransfer;
 
-std::shared_ptr<AbstractContext> privc_ctx();
-std::shared_ptr<OT> ot();
-std::shared_ptr<common::TensorAdapterFactory> tensor_factory();
-std::shared_ptr<TripletGenerator<int64_t, SCALING_N>> tripletor();
-size_t party();
-size_t next_party();
-AbstractNetwork* net();
+template<typename T>
+using OTExtSender = common::OTExtSender<T>;
+template<typename T>
+using OTExtReceiver = common::OTExtReceiver<T>;
+
+static const size_t _g_block_size_expand = sizeof(block) / sizeof(int64_t);
+
+static std::shared_ptr<AbstractContext> privc_ctx() {
+  return paddle::mpc::ContextHolder::mpc_ctx();
+}
+
+static std::shared_ptr<OT> ot() {
+    return std::dynamic_pointer_cast<PrivCContext>(privc_ctx())->ot();
+}
+
+static std::shared_ptr<common::TensorAdapterFactory> tensor_factory() {
+    return paddle::mpc::ContextHolder::tensor_factory();
+}
+
+static std::shared_ptr<TripletGenerator<int64_t, SCALING_N>> tripletor() {
+    return std::dynamic_pointer_cast<PrivCContext>(privc_ctx())->triplet_generator();
+}
+
+static size_t party() {
+    return privc_ctx()->party();
+}
+
+static size_t next_party() {
+    return privc_ctx()->next_party();
+}
+
+static AbstractNetwork* net() {
+  return privc_ctx()->network();
+}
+
 static std::vector<size_t> get_gc_shape(std::vector<size_t> shape,
                                               size_t size = sizeof(int64_t) * 8) {
     // using two int64_t sizes to indicate block size
@@ -47,6 +97,7 @@ static std::vector<size_t> get_block_shape(std::vector<size_t> shape) {
 }
 
 static void block_to_int64(const TensorBlock* input, TensorAdapter<int64_t>* ret) {
+    PADDLE_ENFORCE_EQ(input->numel(), ret->numel() * 2, "input numel no match.");
     const block* in_ptr = reinterpret_cast<const block*>(input->data());
     std::transform(in_ptr, in_ptr + ret->numel(), ret->data(),
                    [](block a) { return *(reinterpret_cast<int64_t*>(&a)); });
