+++ core/privc/privc_context.h
@@ -20,13 +20,12 @@
 #include "core/paddlefl_mpc/mpc_protocol/abstract_network.h"
 #include "core/common/prng.h"
 #include "core/common/rand_utils.h"
-#include "type_utils.h"
 
 namespace privc {
 
 using AbstractNetwork = paddle::mpc::AbstractNetwork;
 using AbstractContext = paddle::mpc::AbstractContext;
-//using block = common::block;
+using block = common::block;
 
 static const size_t SCALING_N = 32;
 
@@ -40,25 +39,12 @@ class PrivCContext : public AbstractContext {
 public:
   PrivCContext(size_t party, std::shared_ptr<AbstractNetwork> network,
                  block seed = common::g_zero_block);
-/*                 block seed = psi::g_zero_block):
-                 AbstractContext::AbstractContext(party, network),
-                 _tripletor{nullptr},
-                 _ot{nullptr} {
-    set_num_party(2);
 
-    if (psi::equals(seed, psi::g_zero_block)) {
-      seed = psi::block_from_dev_urandom();
-    }
-    set_random_seed(seed, 0);
-  }
-*/
   PrivCContext(const PrivCContext &other) = delete;
 
   PrivCContext &operator=(const PrivCContext &other) = delete;
 
   std::shared_ptr<TripletGenerator<int64_t, SCALING_N>> triplet_generator();
-
-  //void set_ot(std::shared_ptr<ObliviousTransfer>& ot);
 
   std::shared_ptr<ObliviousTransfer>& ot();
 
@@ -70,7 +56,6 @@ protected:
 private:
   std::shared_ptr<TripletGenerator<int64_t, SCALING_N>> _tripletor;
   common::PseudorandomNumberGenerator _prng;
-  //psi::PseudorandomNumberGenerator _prng;
   std::shared_ptr<ObliviousTransfer> _ot;
 };
